refactor(svg): extract useSvgGameViewLayout from SvgGameView (refs #75)
Made-with: Cursor
