+++ src/rendering/svg/components/SvgGameView.tsx
@@ -1,4 +1,4 @@
-import { type FC, useCallback, useLayoutEffect, useRef, useState } from 'react';
+import { type FC, useLayoutEffect, useRef } from 'react';
 import { ContainerPoint } from '../../common/ContainerPoint';
 import { SvgBoard } from './SvgBoard';
 import { Game } from '../../../models/Game';
@@ -6,6 +6,7 @@ import { GameHUD } from '../../shell/GameHUD';
 import { ResetViewButton } from '../../shell/ResetViewButton';
 import { useCameraControls } from '../hooks/useCameraControls';
 import { useSvgBoardInteraction } from '../hooks/useSvgBoardInteraction';
+import { useSvgGameViewLayout } from '../hooks/useSvgGameViewLayout';
 import { useWindowLevelGameInput } from '../hooks/useWindowLevelGameInput';
 import { useGameSnapshotBridge } from '../../common/bridge/useGameSnapshotBridge';
 
@@ -23,20 +24,7 @@ export const SvgGameView: FC<SvgGameViewProps> = ({ activeGame, setActiveGame })
     setGameSnapshot
   );
 
-  const [viewSize, setViewSize] = useState({ width: 0, height: 0 });
-
-  const measureContainer = useCallback(() => {
-    const el = containerRef.current;
-    if (!el) return;
-    const r = el.getBoundingClientRect();
-    let w = r.width;
-    let h = r.height;
-    if (w === 0 && h === 0 && typeof window !== 'undefined') {
-      w = window.innerWidth;
-      h = window.innerHeight;
-    }
-    setViewSize({ width: w, height: h });
-  }, []);
+  const { viewSize, measureContainer } = useSvgGameViewLayout(containerRef);
 
   useWindowLevelGameInput({
     onRotateClockwise: cameraPointerCallbacks.onRotateClockwise,
@@ -52,26 +40,6 @@ export const SvgGameView: FC<SvgGameViewProps> = ({ activeGame, setActiveGame })
   useLayoutEffect(() => {
     containerToWorldRef.current = containerToWorld;
   }, [containerToWorld, containerToWorldRef]);
-
-  useLayoutEffect(() => {
-    const el = containerRef.current;
-    if (!el) return;
-
-    const frame = requestAnimationFrame(() => {
-      measureContainer();
-    });
-
-    if (typeof ResizeObserver === 'undefined') {
-      return () => cancelAnimationFrame(frame);
-    }
-
-    const ro = new ResizeObserver(() => measureContainer());
-    ro.observe(el);
-    return () => {
-      cancelAnimationFrame(frame);
-      ro.disconnect();
-    };
-  }, [measureContainer]);
 
   const viewCenter: ContainerPoint = ContainerPoint.xy(viewSize.width / 2, viewSize.height / 2);
